Benchmark large message segment in consumer
A simple benchmark that measures the impact on throughput in the
consumer when the large message segment envelope is used.

This benchmark just sets up a LiKafkaConsumer with a fake consumer
object that always returns the same ConsumerRecords object.

The only difference between the two variants is that one creates all
messages with the large message segment envelope and one just creates
plain old byte[] messages.